# hanghae_node_1/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "hwangsBlog",
        "description": "뭘 좋아할지 몰라서, 다 준비해 봤어요.",
        "version": "0.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "send JWT"
        }
    },
...
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/free/comment": {
            "post": {
                "tags": ["free_comment"],
                "description": "자유 게시판 댓글 작성",
                "parameters": [
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "post_id": {
                                    "example": "초바압"
                                },
                                "content": {
                                    "example": "먹고싶어요!!!"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/free/comment/{post_id}": {
            "get": {
                "tags": ["free_comment"],
                "description": "자유 게시판 댓글 조회",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/free/comment/{comment_id}": {
            "put": {
                "tags": ["free_comment"],
                "description": "자유 게시판 댓글 수정",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "content": {
                                    "example": "오마카세로!!!"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/free\r/comment/{comment_id}": {
            "delete": {
                "tags": ["free_comment"],
                "description": "자유 게시판 댓글 삭제",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/univ/comment": {
            "post": {
                "tags": ["univ_comment"],
                "description": "대학 게시판 댓글 작성",
                "parameters": [
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "post_id": {
                                    "example": "대학댓글초밥"
                                },
                                "content": {
                                    "example": "초밥 지겹니?"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/univ/comment/{post_id}": {
            "get": {
                "tags": ["univ_comment"],
                "description": "대학 게시판 댓글 조회",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/univ/comment/{comment_id}": {
            "put": {
                "tags": ["univ_comment"],
                "description": "대학 게시판 댓글 수정",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "content": {
                                    "example": "나는 안 지겨워"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/univ\r/comment/{comment_id}": {
            "delete": {
                "tags": ["univ_comment"],
                "description": "대학 게시판 댓글 삭제",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}